% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortConceptCount.R
\name{cohortConceptCount}
\alias{cohortConceptCount}
\title{\code{cohortConceptCount()} generates concept_counts}
\usage{
cohortConceptCount(cdm)
}
\arguments{
\item{cdm}{a CDM object}
}
\value{
A CDM table
}
\description{
\code{cohortConceptCount()} generates concept_counts
}
