% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypingTasks.R
\name{cohortOverlap}
\alias{cohortOverlap}
\title{\code{cohortOverlap()} Percentages and counts: Counts only for now, percentages easy. May want to add names to cohorts}
\usage{
cohortOverlap(cdm, cohorts_name)
}
\arguments{
\item{cdm}{A cdm object}

\item{cohorts_name}{Cohort name}
}
\value{
A tibble with a summary of intersections.
}
\description{
\code{cohortOverlap()} Percentages and counts: Counts only for now, percentages easy. May want to add names to cohorts
}
