% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypingTasks.R
\name{largeScaleCharacterization}
\alias{largeScaleCharacterization}
\title{Cohort Characterisation : Large scale + temporal + differneces. Missing differences between them that can be done in shiny step - Also demographics that can be done in previous steps. 8 - Visit Context  ###########
Low priority: tipe of visits Before, during, simultaneous, after
Could potentially be extracted from large scale ?}
\usage{
largeScaleCharacterization(cdm, cohorts_name)
}
\arguments{
\item{cdm}{}

\item{cohorts_name}{}
}
\description{
Cohort Characterisation : Large scale + temporal + differneces. Missing differences between them that can be done in shiny step - Also demographics that can be done in previous steps. 8 - Visit Context  ###########
Low priority: tipe of visits Before, during, simultaneous, after
Could potentially be extracted from large scale ?
}
