% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypingTasks.R
\name{orphanCodes}
\alias{orphanCodes}
\title{\code{orphanCodes()}Details, Cohort Count ,  Cohort definition, Concept Sets, JSON, SQL
TO DO: Ideally separate all steps - more loops but less mess
2,3 - now use the precounted table + Phoebe recommendations
2,3 - We could add options to use getcodeuse from CodelsitGenerator and the Codelist generator search itself
TO DO: Need to improve metadata in names of concept sets and names of cohorts
TO DO: Need to add  Source field and Standard fields}
\usage{
orphanCodes(cdm, cohort_set, concept_counts)
}
\description{
\code{orphanCodes()}Details, Cohort Count ,  Cohort definition, Concept Sets, JSON, SQL
TO DO: Ideally separate all steps - more loops but less mess
2,3 - now use the precounted table + Phoebe recommendations
2,3 - We could add options to use getcodeuse from CodelsitGenerator and the Codelist generator search itself
TO DO: Need to improve metadata in names of concept sets and names of cohorts
TO DO: Need to add  Source field and Standard fields
}
