% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypingTasks.R
\name{timeDistributions}
\alias{timeDistributions}
\title{\code{timeDistributions()} observation time (days) after index , observation time (days) prior to index, time (days) between cohort start and end. Need to add better characterisation of demographics (a sort of table 1)}
\usage{
timeDistributions(cdm, cohorts_name)
}
\arguments{
\item{cdm}{}

\item{cohorts_name}{}
}
\description{
\code{timeDistributions()} observation time (days) after index , observation time (days) prior to index, time (days) between cohort start and end. Need to add better characterisation of demographics (a sort of table 1)
}
